Add files via upload
This recovery was needed by 2 x96 mini users. They installed the MAD rom and then when it rebooted to recovery nothing loaded. They also could not load the recovery manually by pressing the button inside the AV port while giving it power. This recovery worked for both of them. To use it, they formatted a sd card with fat16 or fat32 and put this file on it named recovery.img and then inserted the card without power, began pressing the reset button and then plugged in the power cable while continuing to hold down the reset button until the recovery loaded. Then they continued with the ROM installation instructions