enable pogo autoupdate
Enable autoupdates of pogo:
This checks if installed version is different than latest version. If it is then it checks if the local server configured in pogodroid supports the latest addresses.json file. If it does then it updates pogo.

To disable this, just create the file /sdcard/disableautopogoupdate
it can be empty or have something in it, if it exists we will not update.