adding smali for t95-s1. you dont need this.
tl;dr, this changes nothing for you.

This will not change anything for anybody unless I make a new t95-s1 rom, and then it only changes it for the people who install that rom. The current t95 rom does not have TWRP built in, so it cannot auto-install smali anyways.

I'm looking in to making a final t95 rom with twrp so it can autoload smali.