Pogo changed their packaging. Now we use smart updates
You must update your madmin "system apks" to use this script to update pogo. That also goes for pogo autoupdates as we moved the sanity checks to here and we are going to just tell 16mad to try updating pogo every boot and it will refuse it the version is already right.
Big thanks to Coldshadow and Explodingbanana for doing most of the work here. Also, nice install command Spezus, I used it verbatim :)