you propbably dont need this
there was a small period of time last night where i had a typo in 1.2.9 which i fixed without a version bump, sure enough somebody updated during that time window... this will fix that user. if you're on 1.2.9 and pogodroid updates broke, this should fix it. this probably does not mean you.