remove msg invitation
if somebody else wants to work on trying to recognize google screens the basic idea is you need something unique about each screen as seen here:
  while uiautomator dump -a && sleep 5 && grep -q "com.google.android.gms" /sdcard/window_dump.xml ;do
when uiautomator dump -a is run it makes a file named /sdcard/window_dump.xml. i check it for "com.google.android.gms" to check if the google sign-in is the active app or not.
the actual place where windows are told apart is in the detect_goog_window function
each unique screen which needs to be handled in a unique way needs to be detected uniquely in that /sdcard/window_dump.xml file.